+++ pyvista/core/__init__.py
@@ -6,5 +6,6 @@ from .datasetattributes import DataSetAttributes
 from .filters import (CompositeFilters, DataSetFilters, PolyDataFilters,
                       UnstructuredGridFilters, UniformGridFilters)
 from .grid import Grid, RectilinearGrid, UniformGrid
+from .pyvista_ndarray import pyvista_ndarray
 from .objects import Table, Texture
 from .pointset import PointGrid, PolyData, StructuredGrid, UnstructuredGrid
